+++ code/pilotfile/JSONFileHandler.cpp
@@ -5,7 +5,7 @@ namespace {
 int json_write_callback(const char *buffer, size_t size, void *data) {
 	CFILE* cfp = (CFILE*)data;
 
-	if (cfwrite(buffer, 1, (int)size, cfp) != size) {
+	if ((size_t)cfwrite(buffer, 1, (int)size, cfp) != size) {
 		return -1; // Error
 	} else {
 		return 0; // Success
